requirements.txt
oscpy==0.6.0
pygame==2.1.2
PyOpenGL==3.1.6
